[feat-3] index.js file now changes both ios and android files
